Moving Audit into Domain
