+++ server/domain/category.js
@@ -3,7 +3,7 @@
 var _ = require('underscore');
 var config = require('./../../config/config.json');
 var acl = require('./acl');
-var audit = require('./../audit');
+var audit = require('./audit');
 
 var getCategory = function (id, name, description, columns, features) {
     return {
